+++ config/backend/assets.php
@@ -5,7 +5,7 @@ use MeTools\Core\Plugin;
 return ['Assets' => [
 	'backend-css' => [
 		'input' => [
-			Plugin::path('MeTools').'webroot/css/font-awesome.min',
+			'webroot/vendor/font-awesome/css/font-awesome.min',
 			Plugin::path('MeCms').'webroot/css/backend/bootstrap.min',
 			Plugin::path('MeTools').'webroot/css/default',
 			Plugin::path('MeTools').'webroot/css/forms',
@@ -17,9 +17,9 @@ return ['Assets' => [
 	],
 	'backend-js' => [
 		'input' => [
-			Plugin::path('MeTools').'webroot/js/jquery.min',
+			'webroot/vendor/jquery/jquery.min',
 			Plugin::path('MeCms').'webroot/js/backend/bootstrap.min',
-			Plugin::path('MeCms').'webroot/js/jquery.cookie',
+			'webroot/vendor/jquery-cookie/jquery.cookie',
 			Plugin::path('MeTools').'webroot/js/default',
 			Plugin::path('MeCms').'webroot/js/backend/layout'
 		],