+++ config/me_cms.php
@@ -85,7 +85,12 @@ return ['MeCms' => [
 			//Displays the post category
 			'category' => TRUE,
 			//Displays the post created datetime
-			'created' => TRUE
+			'created' => TRUE,
+			//Displays the Shareaholic social buttons
+			//Remember you have to set app and site IDs. See `shareaholic.app_id` and `shareaholic.site_id`
+			'shareaholic' => FALSE,
+			//Displays the post tags
+			'tags' => TRUE
 		]
 	],
 	//Security
@@ -99,6 +104,15 @@ return ['MeCms' => [
 		'recaptcha' => FALSE,
 		//Interval between searches, in seconds. Set to `0` or `FALSE` to disable
 		'search_interval' => 10
+	],
+	//Shareaholic
+	'shareaholic' => [
+		//App ID. Used for render the "share buttons" of Shareaholic
+		//You can found it on the "Sharing: Edit Share Button Location"
+		'app_id' => '',
+		//Site ID. Used for render the "setup code" of Shareaholic.
+		//You can found it on the "Site Tools Dashboard"
+		'site_id' => ''
 	],
 	//Users
 	'users' => [