+++ src/Console/Installer.php
@@ -24,19 +24,55 @@ namespace MeCms\Console;
 
 use MeTools\Console\Installer as BaseInstaller;
 use Composer\Script\Event;
-use Exception;
 
 /**
  * Provides installation hooks for when this application is installed via
  * composer. Customize this class to suit your needs.
  */
 class Installer extends BaseInstaller {
-    /**
-     * Does some routine installation tasks so people don't have to
-     * @param \Composer\Script\Event $event The composer event object
-	 * @uses MeTools\Console\Installer::postInstall()
+	/**
+	 * Assets for which create symbolic links.
+	 * The key must be relative to `vendor/`, the value must be relative to `webroot/vendor/`
+	 * @see MeTools\Console\Installer::createSymbolicLinkToVendor()
+	 * @var array
 	 */
-    public static function postInstall(Event $event) {
-		parent::postInstall($event);
-    }
+	protected static $linksToAssets = [
+		'components/jquery-cookie'	=> 'jquery-cookie',
+		'newerton/fancy-box/source'	=> 'fancybox',
+		'sunhater/kcfinder'			=> 'kcfinder'
+	];
+	
+	/**
+	 * Fixes Kcfinder.
+	 * Creates the file `vendor/kcfinder/.htaccess`
+	 * @see http://kcfinder.sunhater.com/integrate
+	 */
+	public static function fixKcfinder() {
+		if(!file_exists($file = WWW_ROOT.'vendor'.DS.'kcfinder'.DS.'.htaccess'))
+			(new \Cake\Filesystem\File($file, TRUE))
+				->append('<IfModule mod_php5.c>
+						php_value session.cache_limiter must-revalidate
+						php_value session.cookie_httponly On
+						php_value session.cookie_lifetime 14400
+						php_value session.gc_maxlifetime 14400
+						php_value session.name CAKEPHP
+					</IfModule>');
+	}
+	
+	/**
+	 * Occurs after the autoloader has been dumped, either during install/update, or via the dump-autoload command.
+     * @param \Composer\Script\Event $event The composer event object
+	 * @uses linksToAssets
+	 * @uses MeTools\Console\Installer::linksToAssets
+	 * @uses MeTools\Console\Installer::postAutoloadDump()
+	 */
+	public static function postAutoloadDump(Event $event) {
+		//Merges
+		parent::$linksToAssets = array_merge(parent::$linksToAssets, self::$linksToAssets);
+		
+		parent::postAutoloadDump($event);
+		
+		//Fixes Kcfinder
+		self::fixKcfinder();
+	}
 }