+++ src/Model/Entity/PostsTag.php
@@ -1,0 +1,11 @@
+<?php
+namespace MeCms\Model\Entity;
+
+use Cake\ORM\Entity;
+
+/**
+ * PostsTag entity
+ */
+class PostsTag extends Entity {
+
+}