+++ src/Model/Validation/PostValidator.php
@@ -52,6 +52,60 @@ class PostValidator extends AppValidator {
 		//Text
         $this->requirePresence('text', 'create');
 		
+		//Tag
+        $this->add('tags', [
+			'validTagsLength' => [
+				'message'	=> __d('me_cms', 'Each tag must be between {0} and {1} chars', 3, 20),
+				'rule'		=> [$this, 'validTagsLength']
+			],
+			'validTagsChars' => [
+				'message'	=> sprintf('%s: %s', __d('me_cms', 'Allowed chars'), __d('me_cms', 'lowercase letters, numbers')),
+				'rule'		=> [$this, 'validTagsChars']
+			]
+		]);
+		
         return $this;
+	}
+	
+	/**
+	 * Tags validation method (length).
+	 * For each tag, it checks if the tag has between 3 and 20 chars
+	 * @param string $value Field value
+	 * @param array $context Field context
+	 * @return bool TRUE if is valid, otherwise FALSE
+	 */
+	public function validTagsLength($value, $context) {
+		foreach($value as $tag) {
+			//Continues, if the tag has the ID
+			if(!empty($tag['id']))
+				continue;
+			
+			//Checks if the tag has between 3 and 20 chars
+			if(empty($tag['tag']) || strlen($tag['tag']) < 3 || strlen($tag['tag'] > 20))
+				return FALSE;
+		}
+		
+		return TRUE;
+	}
+	
+	/**
+	 * Tags validation method (chars).
+	 * For each tag, it checks if the tag has only lowercase letters and numbers
+	 * @param string $value Field value
+	 * @param array $context Field context
+	 * @return bool TRUE if is valid, otherwise FALSE
+	 */
+	public function validTagsChars($value, $context) {
+		foreach($value as $tag) {
+			//Continues, if the tag has the ID
+			if(!empty($tag['id']))
+				continue;
+			
+			//Checks if the tag has only lowercase letters and numbers
+			if(empty($tag['tag']) || !(bool) preg_match('/^[a-z0-9]+$/', $tag['tag']))
+				return FALSE;
+		}
+		
+		return TRUE;
 	}
 }