+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # 2.x branch
-## 2.0 branch
+## 2.1 branch
+### 2.1.0-RC2
+* now you can add tags to posts;
+* now you can list posts by date;
+* added support for Shareaholic;
+* jQuery-cookie, Fancybox and KCFinder are installed via Composer;
+* improved the logs management;
+* added error layout and templates;
+* fixed small bugs.
+
 ### 2.0.1-RC1
 * fixed bug, now the cache is flushed automatically, if there's a post-dated post to be published;
 * fixed bug on login with cookie;